requirements.txt
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
fastapi==0.88.0
h11==0.14.0
idna==3.4
kombu==5.2.4
Pillow==9.4.0
prompt-toolkit==3.0.36
pydantic==1.10.4
python-multipart==0.0.5
pytz==2022.7
redis==4.4.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
vine==5.0.0
wcwidth==0.2.5
